[ zen-v2 ] First custom keymap!
That took a while... So many more considerations and edge-cases than I
had anticipated.
I think this strikes a good balance between usable and confusing: most
keys are the same, with numbers and function keys on a different layer.
There's a dedicated layer for Bluetooth, media, and arrow-keys (and the
accents got lumped in there as well). All heavy-lifting is done by the
thumbs, which makes the most sense in my opinion, and the layers are
cycled through using a tap-dance config. I almost wrote NEXT and PREV as
macro functions, but decided against it when I got to conditionals and
name-expanding macros or something arcane like that ^^

Put shift and super/gui on a mod-tap key, before realising that I'd want
to be able to hold _both_ keys when navigating my WM, vim, etc. So more
to tap-dance it was. This is fun!!  : D