[ admin ] Prepare to move Zen config to submodule
Getting ZMK's build to behave nicely with the config in a subdir is more
effort than I can be bothered putting in at the moment, but
unfortunately, I _do_ need my keyboard's firmware to build. So moving to
the intended setup - giving the ZMK config its own repo - and adding
that here as a submodule, seems like the easiest solution for now.